+++ test/utest-online/testlinkapi_online_test.py
@@ -455,6 +455,10 @@ class TestLinkAPIOnlineTestCase(unittest.TestCase):
                             {'cf_b_ex_string' : 'a custom exec value', 
                              'cf_b_ex_numeric' : 111} )
          
+    def test_getExecutionSet_unknownID(self):
+        with self.assertRaisesRegex(TLResponseError, '3000.*40000713'):
+            self.client.getExecutionSet(
+                            '40000713 plan', testcaseexternalid = 'TC-40000712')
         
 if __name__ == "__main__":
     #import sys;sys.argv = ['', 'Test.testName']